requirements.txt
streamlit
scikit-learn==1.3.0
pillow